Handle spaces in baseCommand
SB executor does not put quotes around a
baseCommand element with spaces in it.
The defacto CWL 1.0 interpretation is to
treat this as a quoted string in bash and
not a command.

This modification will result in a baseCommand
that works correctly on both SB executor as
well as cwltool.